HOW TO: change the name of the toolboard during flashing
You have to change the name in all these files, and unplug all the toolboards leaving only the one you want to reflash plugged in.
Change the name written in these files that have changes in this commit.
The flashing will not appear to succeed as it will revert to "waiting for DFU device" after flashing.
Restart the Raspberry Pi and it will appear in /dev/ with the new name.